fix: cli/协议/shared 检视修复（cli --config-dir 透传 + --volume 生效 + pty 断连挂起 + 帧解码错误语义与 64MiB 上限 + FileLoader 副作用/logger 内联）

+++ crates/server/src/services/apps.rs
@@ -392,7 +392,7 @@ pub(crate) async fn handle_apps_ps(msg: Message, state: &Arc<ServerState>) -> Re
                 ProcessStatus::Exited { .. } => "exited".to_string(),
             },
             exit_code: info.status.exit_code(),
-            stdio_len: info.stdio.lock().unwrap().len(),
+            stdio_len: info.stdio.lock().unwrap().len() as u64,
         })
         .collect();
     processes.sort_by_key(|p| p.started_at);
